requirements.txt
streamlit
requests
docarray
inference_client
Pillow
torch
streamlit_image_comparison
